[Refactor] 관리자 계정 따로 분리 #38
