+++ src/main/java/com/example/starhub/service/TechStackService.java
@@ -1,7 +1,5 @@
 package com.example.starhub.service;
 
-import com.example.starhub.dto.request.CreateTechStackRequestDto;
-import com.example.starhub.dto.request.TechStackDto;
 import com.example.starhub.dto.response.TechStackResponseDto;
 import com.example.starhub.entity.TechStackEntity;
 import com.example.starhub.entity.enums.TechCategory;
@@ -10,7 +8,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -35,19 +32,5 @@ public class TechStackService {
                 .map(techStack -> TechStackResponseDto.fromEntity(techStack))
                 .collect(Collectors.toList());
 
-    }
-
-    /**
-     * 기술 스택 생성하기
-     * - 관리자만 허용 가능
-     *
-     * @param techStackDtos 기술 스택 정보 리스트
-     */
-    public void createTechStack(List<TechStackDto> techStackDtos) {
-        List<TechStackEntity> techStackEntities = techStackDtos.stream()
-                                .map(TechStackEntity::createTechStackEntity)
-                                .collect(Collectors.toList());
-
-        techStackRepository.saveAll(techStackEntities); // 여러 개 한 번에 저장
     }
 }
