+++ src/test/java/com/example/starhub/service/TechStackServiceTest.java
@@ -9,13 +9,12 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @SpringBootTest
 @Transactional
@@ -60,53 +59,6 @@ class TechStackServiceTest {
 
         // Then
         assertTrue(techStacks.isEmpty());
-    }
-
-    @Test
-    void createTechStack_shouldSaveTechStacksSuccessfully() {
-        // Given
-        List<TechStackDto> techStackDtos = List.of(
-                buildTechStackDto("Vue", TechCategory.FRONTEND),
-                buildTechStackDto("Django", TechCategory.BACKEND),
-                buildTechStackDto("Swift", TechCategory.MOBILE)
-        );
-
-        // When
-        techStackService.createTechStack(techStackDtos);
-
-        // Then
-        List<TechStackEntity> savedEntities = techStackRepository.findAll();
-        assertEquals(7, savedEntities.size()); // 기존 4개 + 추가 3개
-        assertTrue(savedEntities.stream().anyMatch(stack -> stack.getName().equals("Vue")));
-        assertTrue(savedEntities.stream().anyMatch(stack -> stack.getName().equals("Django")));
-        assertTrue(savedEntities.stream().anyMatch(stack -> stack.getName().equals("Swift")));
-    }
-
-    @Test
-    void createTechStack_shouldHandleEmptyList() {
-        // Given
-        List<TechStackDto> techStackDtos = new ArrayList<>(); // 빈 리스트
-
-        // When
-        techStackService.createTechStack(techStackDtos);
-
-        // Then
-        List<TechStackEntity> savedEntities = techStackRepository.findAll();
-        assertEquals(4, savedEntities.size());
-    }
-
-    @Test
-    void createTechStack_shouldThrowException_whenInvalidDataProvided() {
-        // Given
-        List<TechStackDto> techStackDtos = List.of(
-                buildTechStackDto(null, TechCategory.FRONTEND), // 이름이 null
-                buildTechStackDto("InvalidCategory", null)  // 카테고리가 null
-        );
-
-        // When & Then
-        assertThrows(DataIntegrityViolationException.class, () -> {
-            techStackService.createTechStack(techStackDtos);
-        });
     }
 
     private TechStackEntity saveTechStack(String name, TechCategory category) {
